rename component to Calculator
